requirements.txt
asgiref==3.9.1
Django==5.2.6
django-ckeditor==6.7.3
django-environ==0.12.0
django-js-asset==3.1.2
django-solo==2.4.0
pillow==11.3.0
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1